% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_script_app2.R
\name{analyse_kinetics}
\alias{analyse_kinetics}
\title{Apply functional analysis to an input QFeatures data object}
\usage{
analyse_kinetics(
  data,
  peptide_selection = NULL,
  method = NULL,
  starting_parameters = list(a = NULL, b = 0.001, d = NULL, p = 1),
  formula = NULL
)
}
\arguments{
\item{data}{Input QFeatures data object or a selection}

\item{peptide_selection}{Peptide string chain, e.g., 'DKELKAKGKSAL_4'}

\item{method}{Options:
"fit"  for a regular fitting
"dfit" for differential fitting}

\item{starting_parameters}{To optimise for kinetic model based on 'formula'}

\item{formula}{Functional form or the fitting curve to optimise its parameters given 'starting_parameters'
e.g., 'formula = value ~ a * (1 - exp(-0.07*(timepoint))))'}
}
\value{
Results from Empirical Bayes on fitted kinetic functionals or curves of Deuterium uptake
}
\description{
Apply functional analysis to an input QFeatures data object
}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook
}
