% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_script_app2.R
\name{extract_hdx_data}
\alias{extract_hdx_data}
\title{Extract data provided a filepath}
\usage{
extract_hdx_data(
  data_path,
  normalise = FALSE,
  save = NULL,
  parameters = NULL,
  parameter_file = NULL,
  interactive = FALSE
)
}
\value{
extract data provided an input filepath
}
\description{
Extract data provided a filepath
}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook
}
