% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-class.R
\docType{class}
\name{nlsList-class}
\alias{nlsList-class}
\alias{.nlsList}
\alias{HdxStatModel-class}
\alias{.hdxstatmodel}
\alias{HdxStatModels-class}
\alias{.hdxstatmodels}
\alias{HdxStatRes-class}
\alias{.hdxstatres}
\alias{show,HdxStatModel-method}
\alias{show,HdxStatModels-method}
\alias{show,HdxStatRes-method}
\alias{length,nlsList-method}
\alias{length,HdxStatModel-method}
\alias{length,HdxStatModels-method}
\alias{[[,nlsList,ANY,ANY-method}
\alias{[,nlsList,ANY,ANY,ANY-method}
\title{Classes for hdxstats package}
\usage{
\S4method{show}{HdxStatModel}(object)

\S4method{show}{HdxStatModels}(object)

\S4method{show}{HdxStatRes}(object)

\S4method{length}{nlsList}(x)

\S4method{length}{HdxStatModel}(x)

\S4method{length}{HdxStatModels}(x)

\S4method{[[}{nlsList,ANY,ANY}(x, i, j = "missing", drop = "missing")

\S4method{[}{nlsList,ANY,ANY,ANY}(x, i, j = "missing", drop = "missing")
}
\arguments{
\item{x}{Object to be subset.}

\item{i}{An \code{integer()}. Should be of length 1 for \code{[[}.}

\item{j}{Missing.}

\item{drop}{Missing.}
}
\description{
Classes for hdxstats package
}
\section{Slots}{

\describe{
\item{\code{nlsmodels}}{A \code{list()} containing nls models results in \code{nls} instances.
Each element must be a valid \code{nls} instance.}

\item{\code{nullmodel}}{A \code{nls} object describing the fitted null model}

\item{\code{alternative}}{An \code{nlsList} object described the fitted alternative models}

\item{\code{vis}}{A \code{ggplot} object visualising the fitted objects}

\item{\code{method}}{A \code{character} string indicating the model fitted}

\item{\code{statmodels}}{A \code{list} of models. Each instance must a valid \code{HdxStatModel}.}

\item{\code{results}}{A \code{DataFrame} contains summarised results from an hdx experiment.}

\item{\code{method}}{The statistical method applied method applied.}
}}

