% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalAnalysis-function.R
\name{differentialUptakeKinetics}
\alias{differentialUptakeKinetics}
\alias{computeRSS}
\alias{computeFstat}
\alias{computePval}
\alias{hdxEbayes}
\alias{ttestUptakeKinetics}
\alias{lmUptakeKinetics}
\alias{processFunctional}
\title{This is the main function for fitting functional kinetics to hdx-ms data.
Fitting is performed using the minpack library. By default the model will fit
a Weibull type model but it is possible to provide a formula for more flexibility.
The method using regular expresions to extract the design from the column names
or these can be provided.}
\usage{
differentialUptakeKinetics(
  object,
  feature = NULL,
  design = NULL,
  formula = NULL,
  start = list(a = NULL, b = 0.001, d = NULL, p = 1),
  mycolours = brewer.pal(n = 8, name = "Set2"),
  maxAttempts = 5
)

computeRSS(object)

computeFstat(RSS0, RSS1, d1, d2)

computePval(Fstat, d1, d2)

hdxEbayes(RSS0, RSS1, d1, d2)

ttestUptakeKinetics(object, feature = NULL, design = NULL, formula = NULL)

lmUptakeKinetics(object, feature = NULL, design = NULL, formula = NULL)

processFunctional(object, params)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{feature}{The feature name of which to fit the model}

\item{design}{A character vector indicating replicates, conditions,
chargestates, timepoints ete. Default is NULL and design is extract from column names}

\item{formula}{The formula used. Default is NULL.}

\item{start}{The initial guess for the parameters}

\item{mycolours}{A colour palette default uses \code{brewer.pal}.}

\item{RSS0}{Residual sum of squares for null model}

\item{RSS1}{Residual sum of squares for altenative model}

\item{d1}{The first degrees of freedom for the F-test}

\item{d2}{The second degrees of freedom for the F-test}

\item{Fstat}{The f-statistic}

\item{params}{An object of class \code{HdxStatModels}}
}
\value{
Returns an instance of class \code{HdxStatmodel}

A list containing degrees of freedom and residual sums of squares

A list of the relevent statistics

A p-value according to the F-test

p-values and FDR from empirical Bayes method

Returns an instance of class \code{HdxStatmodel}

Returns an instance of class \code{HdxStatmodel}

An instance of class \code{HdxStatRes} which contains p-values and important
quantities
}
\description{
This is the main function for fitting functional kinetics to hdx-ms data.
Fitting is performed using the minpack library. By default the model will fit
a Weibull type model but it is possible to provide a formula for more flexibility.
The method using regular expresions to extract the design from the column names
or these can be provided.

Computes the residual sum of squares for an \code{HdxStatModel}

Compute the F statistics

Compute p-value based on the F-test

Empirical Bayes computations

This is the main function for using t-tests for hdx-ms data.
The method using regular expresions to extract the design from the column names
or these can be provided.

This is the main function for using linear mixed models for hdx-ms data.
The method using regular expresions to extract the design from the column names
or these can be provided.

This function process the functional fits for hdx-ms data.
}
