% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-methods.R
\name{vcov,HdxStatModel-method}
\alias{vcov,HdxStatModel-method}
\alias{anova,HdxStatModel-method}
\alias{logLik,HdxStatModel-method}
\alias{likRatio,HdxStatModel-method}
\alias{wilk,HdxStatModel-method}
\alias{coef,HdxStatModel-method}
\alias{deviance,HdxStatModel-method}
\alias{residuals,HdxStatModel-method}
\alias{summary,HdxStatModel-method}
\alias{fitUptakeKinetics,QFeatures-method}
\title{Generics for hdxstats package}
\usage{
\S4method{vcov}{HdxStatModel}(object)

\S4method{anova}{HdxStatModel}(object)

\S4method{logLik}{HdxStatModel}(object)

\S4method{likRatio}{HdxStatModel}(object)

\S4method{wilk}{HdxStatModel}(object)

\S4method{coef}{HdxStatModel}(object)

\S4method{deviance}{HdxStatModel}(object)

\S4method{residuals}{HdxStatModel}(object)

\S4method{summary}{HdxStatModel}(object)

\S4method{fitUptakeKinetics}{QFeatures}(
  object,
  feature = NULL,
  design = NULL,
  formula = NULL,
  start = list(a = NULL, b = 0.001, d = NULL, p = 1),
  maxAttempts = 5
)
}
\arguments{
\item{object}{An instance of class \code{QFeatures}}

\item{feature}{The rowname of feature to be modelled}

\item{design}{The design defining conditions and replicates. Default is NULL.}

\item{formula}{The formula for the non-linear fit.}

\item{start}{The initial guess for the parameters. Parameters must match formula}
}
\value{
The variance-covariance matrix of the parameters

An analysis of variance

The log likelihod of the fitted model assuming normally distributed residuals.

The log likelihood ratio between the alternative and null models

Applies Wilk's theorem to generate a p-value based on the likelihood ratio test

Returns the coefficients of the null and alternative models

Returns the deviance of the fitted models

The residuals from the fitted models

Returns a summary of the fitted models.

Returns an instance of \code{HdxStatModels}
}
\description{
Generics for hdxstats package
}
