% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-utils.R
\name{make_parameter_file}
\alias{make_parameter_file}
\alias{preprocess_data}
\alias{extract_hdx_data}
\title{Function to define parameters corresponding to CSV column names to feed into data curation pipeline}
\usage{
make_parameter_file(data, save_parameters = FALSE)

preprocess_data(
  data,
  normalise = FALSE,
  save_qDF = NULL,
  parameters = NULL,
  parameter_file = NULL,
  save_parameters = FALSE,
  interactive = FALSE
)

extract_hdx_data(
  data_path,
  normalise = FALSE,
  save_qDF = NULL,
  parameters = NULL,
  parameter_file = NULL,
  save_parameters = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{data}{Data content from read from CSV file}

\item{save_parameters}{Output path for list with parameters for data curation pipeline}

\item{normalise}{Normalisation method to call \verb{hdxstats::normalisehdx method}}

\item{save_qDF}{Output path for QFeatures object containing curated data}

\item{parameters}{List with parameters for data curation pipeline (OPTIONAL)}

\item{parameter_file}{Path for list with parameters for data curation pipeline (OPTIONAL)}

\item{interactive}{If neither \code{parameters} nor \code{parameter_file} provided, \code{interactive} mode will call \code{make_parameters_file}}

\item{data_path}{CSV data filepath}
}
\value{
list of parameters to CSV column names

QFeatures object

QFeatures object
}
\description{
Function to define parameters corresponding to CSV column names to feed into data curation pipeline

Pre-process data and output a QFeatures instance with curated data from CSV file

Extract data provided a CSV filepath
}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook
}
