% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-utils.R
\name{exchangeableAmides}
\alias{exchangeableAmides}
\alias{normalisehdx}
\title{Compute exchangeable amides.}
\usage{
exchangeableAmides(sequence)

normalisehdx(object, sequences = NULL, method = "pc", correction = NULL)
}
\arguments{
\item{sequence}{The sequence of the peptide}

\item{object}{An object of class QFeatures}

\item{sequences}{Sequences corresponding to the hdxdata rows}

\item{method}{Either "pc" or "bc". Percentgae incoporation or
back-exchange correction. Default is "pc"}

\item{correction}{A numeric vector indicated complete incoperation values
for correction. Must be provided if method is "bc"}
}
\value{
Returns a numeric indicating the number of exchangeable amides

Returns a normalised 'QFeatures' object.
}
\description{
Computes the number of exchangeable amides based on the sequnece

Normalise hdx data by putting on percentage incoperation scale or
using back-exchange correction
}
