% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-generics.R
\name{likRatio}
\alias{likRatio}
\alias{wilk}
\alias{fitUptakeKinetics}
\title{Generics for hdxstats package}
\usage{
likRatio(object, ...)

wilk(object, ...)

fitUptakeKinetics(object, ...)
}
\arguments{
\item{object}{An object of class \code{HdxStatModel}}

\item{...}{Additional arguments to \code{likRatio}}
}
\description{
Generics for hdxstats package
}
