% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-function.R
\name{parseDeutData}
\alias{parseDeutData}
\title{Parse deuterium data into QFeature objects}
\usage{
parseDeutData(
  object,
  assayname = "incoperation",
  rownames = NULL,
  quantcol = NULL,
  filter = NULL,
  filterScore = 0.9,
  sequence = "pep_sequence",
  charge = "pep_charge",
  design = NULL
)
}
\arguments{
\item{object}{An object of class DataFrame}

\item{assayname}{The name of assay. The Default is incoperation.}

\item{rownames}{The rownames of the features. Default is NULL, in which case
the rownames are extract from the \code{sequence} and \code{charge} in the object}

\item{quantcol}{The columns wiht the quantitative data.}

\item{filter}{Currently unused filter}

\item{filterScore}{Currently usused filter score}

\item{sequence}{The name of the column where the peptide sequence is stored.
Default is "pep_sequence"}

\item{charge}{The name of the column where the peptide charge is stored.
Default is "pep_chrage"}

\item{design}{The design which will become the column names}
}
\value{
An instance of class \code{QFeatures} storing the quantitative mass-spectrometry
data
}
\description{
Parse deuterium data into QFeature objects
}
