% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb-visualisation.R
\name{define_color_function}
\alias{define_color_function}
\alias{hdx_to_pdb_colours}
\title{Color function}
\usage{
define_color_function(dataset, scale_limits = NULL, cmap_name = NULL)

hdx_to_pdb_colours(
  dataset,
  pdb_filepath,
  scale_limits = NULL,
  cmap_name = NULL
)
}
\arguments{
\item{dataset}{A list of estimated protection (positive) and deprotection (negative) values per residue}

\item{scale_limits}{You can force a range of numerical values to be mapped only}

\item{pdb_filepath}{The path to the PDB file for mapping of (de)protection values}
}
\value{
Returns a color function

Returns a list of colours and residue numbers to be inputted into NGLVieweR,
with protection values in Red and deprotection values in Blue, and NA values in Grey.
}
\description{
Color function

Map estimated (de)protection values to Blue-White-Red colormap values
}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook

Broncio Aguilar-Sanjuan and Oliver Crook
}
