% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_script_app2.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process data and output a QFeatures instance}
\usage{
preprocess_data(
  data,
  normalise = FALSE,
  save = NULL,
  parameters = NULL,
  parameter_file = NULL,
  interactive = FALSE
)
}
\value{
preprocessed data
}
\description{
Pre-process data and output a QFeatures instance
}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook
}
