% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-streamline.R
\name{visualise_hdx_data}
\alias{visualise_hdx_data}
\title{Extract data provided a CSV filepath}
\usage{
visualise_hdx_data(
  results,
  data_selection = NULL,
  type = NULL,
  level = NULL,
  fasta = NULL,
  pdb = NULL
)
}
\arguments{
\item{results}{List of functional analysis results}

\item{data_selection}{QFeatures object selection}

\item{type}{Visualisation type for functional analysis results
Options:
'kinetics' Visualise fittings to Deu uptake kinetics data
'forest' Visualise quality of fitting parameters for model
'manhattan' Visualise ...
'epitope' Visualise epitope map to see data overlap
'protection' Visualise protection/deprotection plots}

\item{level}{Options:
'peptide' applicable to 'type = epitope'
'residue' applicable to 'type = epitope' and 'type = protection'}

\item{fasta}{Path to FASTA file containing the sequence of protein of interest}

\item{pdb}{Path to PDB coordinate file of protein of interest}
}
\value{
ggplot object(s) or NGLVieweR object (if pdb not NULL)
}
\description{
Extract data provided a CSV filepath
}
\author{
Broncio Aguilar-Sanjuan and Oliver Crook
}
