% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalAnalysis-plots.R
\name{hmpWindow}
\alias{hmpWindow}
\alias{manhattanplot}
\alias{plotEpitopeMap}
\alias{plotEpitopeMapFdr}
\alias{plotEpitopeMapResidue}
\alias{ComputeAverageMap}
\alias{plotAverageMaps}
\alias{forestPlot}
\alias{hdxdifference}
\alias{hdxheatmap}
\title{hmpWindow (Currently in development and not used)}
\usage{
hmpWindow(params, sequences, region, interval = 20)

manhattanplot(params, sequences, difference = 0, region = NULL, nrow = 1)

plotEpitopeMap(
  AAString,
  peptideSeqs,
  numlines = 5,
  maxmismatch = 0,
  by = 5,
  scores = NULL,
  name = "-log10 p values",
  threshold = -log10(0.05)
)

plotEpitopeMapFdr(
  AAString,
  peptideSeqs,
  numlines = 5,
  maxmismatch = 0,
  by = 5,
  scores = NULL,
  name = "-log10 p values",
  threshold = -log10(0.05)
)

plotEpitopeMapResidue(
  AAString,
  peptideSeqs,
  numlines = 5,
  maxmismatch = 0,
  by = 5,
  scores = NULL,
  name = "-log10 p values",
  threshold = -log10(0.05)
)

ComputeAverageMap(
  AAString,
  peptideSeqs,
  numlines = 5,
  maxmismatch = 0,
  by = 5,
  scores = NULL,
  name = "-log10 p values",
  threshold = -log10(0.05)
)

plotAverageMaps(averageMaps, name = "-log10 p value", numlines = 2, by = 5)

forestPlot(params, condition = c(1, 2))

hdxdifference(
  object,
  AAString,
  peptideSeqs,
  numlines = 5,
  maxmismatch = 0,
  by = 5,
  cols = c(1, 4),
  scores = NULL,
  name = "-log10 p value"
)

hdxheatmap(
  averageMaps,
  diffMaps,
  name = "-log10 p value (signed)",
  numlines = 2,
  by = 5
)
}
\arguments{
\item{params}{An object of class \code{HdxStatRes}}

\item{sequences}{A character vector containing the measured peptides sequences,
if there are different charge states the vector must identify the charge state.}

\item{region}{The start and end of the sequences provided. Columns must be
called "Start" and "End".}

\item{interval}{interval size to average over.}

\item{difference}{A numeric vector with deuterium differences for each peptide}

\item{nrow}{The number of rows to plot the manhatten plot over. Useful for larger
proteins}

\item{AAString}{An object of class \code{AAString} for the protein of interest}

\item{peptideSeqs}{A character vector of peptide sequences}

\item{numlines}{The number of lines to plot the protein over. Useful for larger
proteins. Default is 2.}

\item{maxmismatch}{A numeric indicating if incorrect mapping is allowed. Number
indicated the number of mismatched amino acids. Default is 0.}

\item{by}{A value to indicate the legend breaks. Default is NULL.}

\item{scores}{A numeric vector indicating score to be used for plotting. Most
likely adjusted p-values.}

\item{name}{The name of the legend for the score plotting.
Default is "-log10 p values (singed)". Indicated the significance and direction}

\item{threshold}{The threshold used to determine significance. Default is
\code{-log10(0.05)}. Note the log scale.}

\item{averageMaps}{A list of average maps generated by the \code{computeAverageMaps}
function}

\item{condition}{If there are multiple conditions which ones to plot. Default
is \code{c(1,2)}}

\item{object}{An object of class \code{QFeatures} contains the hdx data.}

\item{cols}{Columns for which to compute the difference. The difference is
relative to the first entry.}

\item{diffMaps}{A list of difference maps generated by the \code{hdxdifference}
function}
}
\value{
use for side effect which returns manhattan plot

use for side effect which returns manhatten plot

Side effect produces a forest plot
}
\description{
A manhatten plot for a multi-level testing prodcedure based on the
haromonic mean p-value

A manhattan plot for epitope mapping

Generate an epitope map by plotting the signifcantly changed peptides
with respect to a differential HDX-MS experiment
}
