% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdxstat-generics.R
\name{likRatio}
\alias{likRatio}
\alias{wilk}
\alias{fitUptakeKinetics}
\title{Generics for hdxstats package}
\usage{
likRatio(object, ...)

wilk(object, ...)

fitUptakeKinetics(object, ...)
}
\description{
Generics for hdxstats package
}
